+++ __test__/DownloadURLFactory.test.ts
@@ -1,10 +1,14 @@
-import { OS, Arch } from "../src/platform";
-import { DownloadURLFactory } from "../src/DownloadURLFactory";
+import { describe, expect, test } from "vitest";
 import { ArchiveDownloadURL, LatestDownloadURL } from "../src/DownloadURL";
+import { DownloadURLFactory } from "../src/DownloadURLFactory";
+import { Arch, OS } from "../src/platform";
 
 describe("DownloadURLFactory", () => {
   describe.each([
     ["80.0", ArchiveDownloadURL],
+    ["firefox-80.0", ArchiveDownloadURL],
+    ["beta-80.0b1", ArchiveDownloadURL],
+    ["devedition-80.0b1", ArchiveDownloadURL],
     ["latest", LatestDownloadURL],
     ["latest-beta", LatestDownloadURL],
     ["latest-devedition", LatestDownloadURL],